+++ src/Api/ResourceNotFoundException.php
@@ -25,17 +25,16 @@ class ResourceNotFoundException extends RuntimeException
     /**
      * Creates a new exception for a resource path.
      *
-     * @param string    $path     The path which was not found.
-     * @param int       $code     The error code.
-     * @param Exception $previous The exception that caused this exception.
+     * @param string    $path  The path which was not found.
+     * @param Exception $cause The exception that caused this exception.
      *
      * @return static The created exception.
      */
-    public static function forPath($path, $code = 0, Exception $previous = null)
+    public static function forPath($path, Exception $cause = null)
     {
         return new static(sprintf(
             'The resource %s does not exist.',
             $path
-        ), $code, $previous);
+        ), 0, $cause);
     }
 }